Move alias message into alias class
